requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
snowflake-connector-python==3.4.0
pandas==2.1.4
pytest==7.4.3

